requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==0.7.1            # via tensorboard, tensorflow
ansimarkup==1.4.0         # via better-exceptions-fork, loguru
appdirs==1.4.3            # via black, dvc
asciimatics==1.11.0       # via dvc
astor==0.8.0              # via tensorflow
attrs==19.1.0             # via black
better-exceptions-fork==0.2.1.post6  # via loguru
black==19.3b0
boto3==1.9.170            # via pytorch-pretrained-bert
botocore==1.12.170        # via boto3, s3transfer
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via dulwich, requests
chardet==3.0.4            # via requests
click==7.0                # via black, pip-tools
colorama==0.4.1           # via ansimarkup, dvc, loguru
configobj==5.0.6          # via dvc
configparser==3.7.4       # via dvc
contextlib2==0.5.5        # via schema
decorator==4.4.0          # via jsonpath-ng, networkx
distro==1.4.0             # via dvc
docutils==0.14            # via botocore
dulwich==0.19.11          # via dvc
dvc[gs]==0.41.3
future==0.17.1            # via asciimatics, dvc, grandalf
gast==0.2.2               # via tensorflow
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11         # via dvc
google-api-core==1.12.0   # via google-cloud-core, google-cloud-storage
google-auth==1.6.3        # via google-api-core
google-cloud-core==0.28.1  # via google-cloud-storage
google-cloud-storage==1.13.0  # via dvc
google-pasta==0.1.7       # via tensorflow
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos==1.6.0  # via google-api-core
grandalf==0.6             # via dvc
grpcio==1.21.1            # via tensorboard, tensorflow
h5py==2.9.0               # via keras-applications
humanize==0.5.1           # via dvc
idna==2.8                 # via requests
inflect==2.1.0            # via dvc
jmespath==0.9.4           # via boto3, botocore
joblib==0.13.2            # via scikit-learn
json-tricks==3.13.1
jsonpath-ng==1.4.3        # via dvc
keras-applications==1.0.8  # via tensorflow
keras-preprocessing==1.1.0
loguru==0.2.5
markdown==3.1.1           # via tensorboard
nanotime==0.5.2           # via dvc
networkx==2.3             # via dvc
numpy==1.16.4             # via h5py, keras-applications, keras-preprocessing, pandas, pytorch-nlp, pytorch-pretrained-bert, scikit-learn, scipy, tensorboard, tensorflow, torch
pandas==0.24.2            # via pytorch-nlp
pillow==6.0.0             # via asciimatics
pip-tools==3.8.0
ply==3.11                 # via dvc, jsonpath-ng
protobuf==3.8.0           # via google-api-core, googleapis-common-protos, tensorboard, tensorflow
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via dvc, pyasn1-modules, rsa
pyfiglet==0.8.post1       # via asciimatics
pygments==2.4.2           # via better-exceptions-fork
pyparsing==2.4.0          # via grandalf
python-dateutil==2.8.0    # via botocore, pandas
pytorch-nlp==0.4.1
pytorch-pretrained-bert==0.6.2
pytz==2019.1              # via google-api-core, pandas
regex==2019.6.8           # via pytorch-pretrained-bert
requests==2.22.0          # via dvc, google-api-core, pytorch-nlp, pytorch-pretrained-bert
rsa==4.0                  # via google-auth
ruamel.yaml==0.15.97      # via dvc
s3transfer==0.2.1         # via boto3
schema==0.7.0             # via dvc
scikit-learn==0.21.2
scipy==1.3.0              # via scikit-learn
six==1.12.0               # via absl-py, configobj, google-api-core, google-auth, google-resumable-media, grpcio, h5py, jsonpath-ng, keras-preprocessing, pip-tools, protobuf, python-dateutil, tensorboard, tensorflow
smmap2==2.0.5             # via gitdb2
tensorboard==1.14.0       # via tensorflow
tensorflow-estimator==1.14.0rc1  # via tensorflow
tensorflow==1.14.0
termcolor==1.1.0          # via tensorflow
toml==0.10.0              # via black
torch==1.1.0              # via pytorch-pretrained-bert
tqdm==4.32.1              # via pytorch-nlp, pytorch-pretrained-bert
treelib==1.5.5            # via dvc
urllib3==1.25.3           # via botocore, dulwich, requests
wcwidth==0.1.7            # via asciimatics
werkzeug==0.15.4          # via tensorboard
wheel==0.33.4             # via tensorboard, tensorflow
wrapt==1.11.2             # via tensorflow
zc.lockfile==1.4          # via dvc

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via dvc, google-api-core, markdown, protobuf, tensorboard, zc.lockfile
